UI improvements round 1
