+++ plugins/emergencyMultisig/pages/new.tsx
@@ -1,166 +1,63 @@
-import React, { ReactNode, useEffect, useState } from "react";
+import React from "react";
 import { Button, IconType, Icon, InputText, TextAreaRichText } from "@aragon/ods";
-import { useAccount, useWaitForTransactionReceipt, useWriteContract } from "wagmi";
-import { toHex } from "viem";
-import { useAlerts } from "@/context/Alerts";
+import { useAccount } from "wagmi";
 import WithdrawalInput from "@/components/input/withdrawal";
 import { FunctionCallForm } from "@/components/input/function-call-form";
-import { ProposalMetadata, RawAction } from "@/utils/types";
-import { useRouter } from "next/router";
 import { Else, ElseIf, If, Then } from "@/components/if";
 import { PleaseWaitSpinner } from "@/components/please-wait";
-import {
-  PUB_DUAL_GOVERNANCE_PLUGIN_ADDRESS,
-  PUB_CHAIN,
-  PUB_EMERGENCY_MULTISIG_PLUGIN_ADDRESS,
-  PUB_APP_NAME,
-  PUB_PROJECT_URL,
-} from "@/constants";
 import { ActionCard } from "@/components/actions/action";
-import { uploadToPinata } from "@/utils/ipfs";
-import { EmergencyMultisigPluginAbi } from "../artifacts/EmergencyMultisigPlugin";
-import { usePublicKeyRegistry } from "../hooks/usePublicKeyRegistry";
-import { useEncryptedData } from "../hooks/useEncryptedData";
-import { MissingContentView } from "../components/MissingContentView";
+import { MissingContentView } from "@/components/MissingContentView";
 import { useWeb3Modal } from "@web3modal/wagmi/react";
 import { useDerivedWallet } from "../hooks/useDerivedWallet";
 import { MainSection } from "@/components/layout/main-section";
-
-enum ActionType {
-  Signaling,
-  Withdrawal,
-  Custom,
-}
+import { useCreateProposal } from "../hooks/useCreateProposal";
+import { useCanCreateProposal } from "../hooks/useCanCreateProposal";
+import { ActionType } from "@/utils/types";
+import { usePublicKeyRegistry } from "../hooks/usePublicKeyRegistry";
 
 export default function Create() {
-  const { push } = useRouter();
+  const { open } = useWeb3Modal();
   const { address: selfAddress, isConnected } = useAccount();
   const { publicKey, requestSignature } = useDerivedWallet();
-  const { open } = useWeb3Modal();
-  const { addAlert } = useAlerts();
-  const [title, setTitle] = useState<string>("");
-  const [summary, setSummary] = useState<string>("");
-  const [description, setDescription] = useState<string>("");
-  const [actions, setActions] = useState<RawAction[]>([]);
-  const { writeContract: createProposalWrite, data: createTxHash, error, status } = useWriteContract();
-  const { isLoading: isConfirming, isSuccess: isConfirmed } = useWaitForTransactionReceipt({ hash: createTxHash });
-  const [actionType, setActionType] = useState<ActionType>(ActionType.Signaling);
+  const canCreate = useCanCreateProposal();
+  const {
+    title,
+    summary,
+    description,
+    actions,
+    actionType,
+    setTitle,
+    setSummary,
+    setDescription,
+    setActions,
+    setActionType,
+    isCreating,
+    submitProposal,
+  } = useCreateProposal();
   const {
     data: { addresses: registeredSigners },
     registerPublicKey,
-    isLoading: pubKeysLoading,
+    isLoading: isLoadingPubKeys,
+    isConfirming: isRegisteringPublicKey,
   } = usePublicKeyRegistry();
-  const { encryptProposalData } = useEncryptedData();
 
   const changeActionType = (actionType: ActionType) => {
     setActions([]);
     setActionType(actionType);
   };
-
-  useEffect(() => {
-    if (status === "idle" || status === "pending") return;
-    else if (status === "error") {
-      if (error?.message?.startsWith("User rejected the request")) {
-        addAlert("Transaction rejected by the user", {
-          timeout: 4 * 1000,
-        });
-      } else {
-        console.error(error);
-        addAlert("Could not create the proposal", { type: "error" });
-      }
-      return;
-    }
-
-    // success
-    if (!createTxHash) return;
-    else if (isConfirming) {
-      addAlert("Proposal submitted", {
-        description: "Waiting for the transaction to be validated",
-        txHash: createTxHash,
-      });
-      return;
-    } else if (!isConfirmed) return;
-
-    addAlert("Proposal created", {
-      description: "The transaction has been validated",
-      type: "success",
-      txHash: createTxHash,
-    });
-    setTimeout(() => push("#/"), 1000 * 2);
-  }, [status, createTxHash, isConfirming, isConfirmed]);
-
-  const submitProposal = async () => {
-    // Check metadata
-    if (!title.trim())
-      return addAlert("Invalid proposal details", {
-        description: "Please, enter a title",
-        type: "error",
-      });
-
-    if (!summary.trim())
-      return addAlert("Invalid proposal details", {
-        description: "Please, enter a summary of what the proposal is about",
-        type: "error",
-      });
-
-    // Check the action
-    switch (actionType) {
-      case ActionType.Signaling:
-        break;
-      case ActionType.Withdrawal:
-        if (!actions.length) {
-          return addAlert("Invalid proposal details", {
-            description: "Please ensure that the withdrawal address and the amount to transfer are valid",
-            type: "error",
-          });
-        }
-        break;
-      default:
-        if (!actions.length || !actions[0].data || actions[0].data === "0x") {
-          return addAlert("Invalid proposal details", {
-            description: "Please ensure that the values of the action to execute are complete and correct",
-            type: "error",
-          });
-        }
-    }
-
-    const privateMetadata: ProposalMetadata = {
-      title,
-      summary,
-      description,
-      resources: [{ name: PUB_APP_NAME, url: PUB_PROJECT_URL }],
-    };
-
-    // Encrypt the proposal data
-    const { payload: publicMetadataJson, hashed } = await encryptProposalData(privateMetadata, actions);
-
-    const publicMetadataUri = await uploadToPinata(JSON.stringify(publicMetadataJson));
-
-    createProposalWrite({
-      chainId: PUB_CHAIN.id,
-      abi: EmergencyMultisigPluginAbi,
-      address: PUB_EMERGENCY_MULTISIG_PLUGIN_ADDRESS,
-      functionName: "createProposal",
-      args: [toHex(publicMetadataUri), hashed.metadataUri, hashed.actions, PUB_DUAL_GOVERNANCE_PLUGIN_ADDRESS, false],
-    });
-  };
-
   const handleTitleInput = (event: React.ChangeEvent<HTMLInputElement>) => {
     setTitle(event?.target?.value);
   };
-
   const handleSummaryInput = (event: React.ChangeEvent<HTMLInputElement>) => {
     setSummary(event?.target?.value);
   };
-
-  const showLoading = status === "pending" || isConfirming;
 
   return (
     <MainSection className="flex flex-col gap-y-6 md:px-16 md:py-10">
       <div className="w-full justify-between">
         <h1 className="mb-6 text-3xl font-semibold text-neutral-900">Create Proposal</h1>
 
-        <If condition={pubKeysLoading}>
+        <If condition={isLoadingPubKeys}>
           <Then>
             {/* No public keys yet */}
             <div>
@@ -183,6 +80,7 @@ export default function Create() {
                   generate an encryption key only for this DAO.`}
               callToAction="Register your public key"
               onClick={() => registerPublicKey()}
+              isLoading={isRegisteringPublicKey}
             />
           </ElseIf>
           <ElseIf condition={!publicKey}>
@@ -191,6 +89,12 @@ export default function Create() {
               message={`Please, sign in with your wallet in order to decrypt the private proposal data.`}
               callToAction="Sign in to continue"
               onClick={() => requestSignature()}
+            />
+          </ElseIf>
+          <ElseIf condition={!canCreate}>
+            {/* Not a member */}
+            <MissingContentView
+              message={`You cannot create proposals on the multisig because you are not currently defined as a member.`}
             />
           </ElseIf>
           <Else>
@@ -294,17 +198,18 @@ export default function Create() {
               </span>
             </div>
 
-            <If condition={showLoading}>
+            <If condition={actionType !== ActionType.Custom}>
               <Then>
-                <div className="mb-6 mt-14">
-                  <PleaseWaitSpinner fullMessage="Confirming transaction..." />
-                </div>
-              </Then>
-              <ElseIf condition={actionType !== ActionType.Custom}>
-                <Button className="mb-6 mt-14" size="lg" variant="primary" onClick={() => submitProposal()}>
+                <Button
+                  isLoading={isCreating}
+                  className="mb-6 mt-14"
+                  size="lg"
+                  variant="primary"
+                  onClick={() => submitProposal()}
+                >
                   Submit proposal
                 </Button>
-              </ElseIf>
+              </Then>
               <Else>
                 <div className="mb-6 mt-14">
                   <If not={actions.length}>
