+++ plugins/emergencyMultisig/pages/proposal-list.tsx
@@ -16,7 +16,7 @@ import Link from "next/link";
 import { Else, ElseIf, If, Then } from "@/components/if";
 import { PUB_EMERGENCY_MULTISIG_PLUGIN_ADDRESS, PUB_CHAIN } from "@/constants";
 import { useDerivedWallet } from "../hooks/useDerivedWallet";
-import { MissingContentView } from "../components/MissingContentView";
+import { MissingContentView } from "@/components/MissingContentView";
 import { MainSection } from "@/components/layout/main-section";
 
 const DEFAULT_PAGE_SIZE = 6;
